Move eigen dl version to single file, update to 3.3.2
This moves the eigen download version and extraction path to a single
file (.eigen-build-version), loaded by both the appveyor and travis-ci
scripts.

It also updates the eigen version from 3.3.0 to the latest 3.3.2.

Finally it saves a tiny amount of disk space on the travis-ci builds by
extracting the .tar.gz on the fly rather than downloading first then
decompressing the saved file.