Make slice::compute return signed integers
Signed results are necessary for at least the step, which may
be negative.

If unsigned results are needed for the others, this should be
done with a static_cast; the result of a reintepret_cast
between signed and unsigned is compiler-defined.  I think it's
better to just pass the ssize_t values directly to the user
and let them handle a conversion to unsigned if needed.