Fix compilation with Clang and C++14
It's not clear why the other changes related to arrays triggered this
failure, but for some reason the compiler is generating references to
the symbol for the digits member of the int_to_str type.

Added the definition of the digits member so that the symbol will be
emitted if needed. A quick check suggests that this doesn't increase the
compiled size at all. I'm slightly suspicious because the size appears
to stay *exactly* the same, whereas it should go up just to make room
for the definition in the symbol table, but maybe some alignment is at
work.