def_readwrite/readonly: bind using class type, not inferred type
Current def_readwrite/def_readonly create getter/setter lambda that
take the inferred type `C` as the first argument, but this is a problem
if the member is actually a base class member (e.g.

    derived.def_readonly("x", &Base::x)

because this ends up creating a setter/getter that aren't callable with
a derived instance when `Base` isn't registered: it invariably produces
an invalid argument failure when attempting to access the property due
to the `D` instance/`B` argument mismatch.

This commit changes the generated lambdas to use the `type` (as given in
`class_<type>`) rather than the inferred `C` for the lambdas, and adds a
static_assert failure for an attempt to call it with a member that isn't
part of `type` (or a base class).